# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Knissa API
  version: 1.0.0
  description: Backend API for the Knissa Financial Platform.

  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Local Development Server
...
paths:
...
  /exchange-rates:
    post:
      tags:
        - Exchange Rates

      summary: Create exchange rate
      operationId: createExchangeRate

      security:
        - BearerAuth: []

      requestBody:
        required: true

        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExchangeRateRequest"

      responses:
        "201":
          description: Exchange rate created.

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeRateResponse"

        "400":
          description: Validation error

        "401":
          description: Unauthorized

    get:
      tags:
        - Exchange Rates

      summary: List exchange rates
      operationId: listExchangeRates

      security:
        - BearerAuth: []

      responses:
        "200":
          description: List of exchange rates.

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeRateListResponse"

        "401":
          description: Unauthorized
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ExchangeRateRequest:
      type: object

      required:
        - fromCurrency
        - toCurrency
        - rate

      properties:
        fromCurrency:
          type: string
          example: BRL

        toCurrency:
          type: string
          example: HTG

        rate:
          type: number
          format: double
          example: 24.50
    ExchangeRate:
      type: object

      properties:
        id:
          type: string
          format: uuid

        fromCurrency:
          type: string

        toCurrency:
          type: string

        rate:
          type: number
          format: double

        createdAt:
          type: string
          format: date-time
    ExchangeRateResponse:
      type: object

      properties:
        success:
          type: boolean
          example: true

        data:
          $ref: "#/components/schemas/ExchangeRate"
    ExchangeRateListResponse:
      type: object

      properties:
        success:
          type: boolean
          example: true

        data:
          type: array

          items:
            $ref: "#/components/schemas/ExchangeRate"
...